routing: queues: improve remote queues
Improve the code quality of the `InternalRemoteQueue` by splitting the
remote trigger service queue and live data service queue. The trigger
service queue is now implemented in the `RemoteTriggerQueue` class.
