+++ SensateIoT.Platform.Router.Common/Services/Metrics/HealthMonitoringService.cs
@@ -18,10 +18,12 @@ namespace SensateIoT.Platform.Router.Common.Services.Metrics
 		private readonly HealthCheckSettings m_settings;
 		private readonly IPublicMqttClient m_publicMqttClient;
 		private readonly IInternalMqttClient m_internalMqttClient;
+		private readonly IRemoteTriggerQueue m_triggerQueue;
 
 		public bool IsHealthy => this.GetHealthStatus();
 
 		public HealthMonitoringService(IQueue<IPlatformMessage> inputQueue,
+									   IRemoteTriggerQueue triggerQueue,
 									   IInternalRemoteQueue @internal,
 									   IPublicRemoteQueue @public,
 									   IPublicMqttClient publicClient,
@@ -33,6 +35,7 @@ namespace SensateIoT.Platform.Router.Common.Services.Metrics
 			this.m_inputQueue = inputQueue;
 			this.m_internalMqttClient = internalClient;
 			this.m_publicMqttClient = publicClient;
+			this.m_triggerQueue = triggerQueue;
 			this.m_settings = settings.Value;
 		}
 
@@ -97,13 +100,13 @@ namespace SensateIoT.Platform.Router.Common.Services.Metrics
 		private bool CheckLiveDataQueues()
 		{
 			var limit = this.m_settings.LiveDataServiceQueueLimit ?? this.m_settings.DefaultQueueLimit;
-			return this.m_internalRemoteQueues.LiveDataQueueLength <= limit;
+			return this.m_internalRemoteQueues.Count <= limit;
 		}
 
 		private bool CheckTriggerQueues()
 		{
 			var limit = this.m_settings.TriggerServiceQueueLimit ?? this.m_settings.DefaultQueueLimit;
-			return this.m_internalRemoteQueues.TriggerQueueLength <= limit;
+			return this.m_triggerQueue.Count <= limit;
 		}
 	}
 }
