+++ SensateIoT.Platform.Router.Common/Services/Processing/RoutingPublishService.cs
@@ -8,8 +8,10 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+
 using SensateIoT.Platform.Router.Common.Collections.Abstract;
 using SensateIoT.Platform.Router.Common.Services.Background;
 using SensateIoT.Platform.Router.Common.Settings;
@@ -19,10 +21,12 @@ namespace SensateIoT.Platform.Router.Common.Services.Processing
 	public class RoutingPublishService : TimedBackgroundService
 	{
 		private readonly IInternalRemoteQueue m_internalRemote;
+		private readonly IRemoteTriggerQueue m_triggerQueue;
 		private readonly IRemoteStorageQueue m_remoteStorageQueue;
 		private readonly IRemoteNetworkEventQueue m_eventQueue;
 
 		public RoutingPublishService(IInternalRemoteQueue internalRemote,
+									 IRemoteTriggerQueue triggerQueue,
 									 IRemoteStorageQueue remoteStorage,
 									 IRemoteNetworkEventQueue remoteEvents,
 									 IOptions<RoutingQueueSettings> options,
@@ -31,12 +35,13 @@ namespace SensateIoT.Platform.Router.Common.Services.Processing
 			this.m_internalRemote = internalRemote;
 			this.m_remoteStorageQueue = remoteStorage;
 			this.m_eventQueue = remoteEvents;
+			this.m_triggerQueue = triggerQueue;
 		}
 
 		protected override async Task ExecuteAsync(CancellationToken token)
 		{
 			await Task.WhenAll(this.m_internalRemote.FlushAsync(),
-							   this.m_internalRemote.FlushLiveDataAsync(),
+							   this.m_triggerQueue.FlushAsync(),
 							   this.m_eventQueue.FlushEventsAsync(),
 							   this.m_remoteStorageQueue.FlushMessagesAsync()).ConfigureAwait(false);
 		}
