+++ SensateIoT.Platform.Router.Service/Init/RoutingInitExtensions.cs
@@ -48,6 +48,7 @@ namespace SensateIoT.Platform.Router.Service.Init
 			services.AddSingleton<IQueue<IPlatformMessage>, MessageQueue>();
 			services.AddSingleton<IRemoteNetworkEventQueue, RemoteNetworkEventQueue>();
 			services.AddSingleton<IInternalRemoteQueue, InternalMqttQueue>();
+			services.AddSingleton<IRemoteTriggerQueue, RemoteTriggerQueue>();
 			services.AddSingleton<IPublicRemoteQueue, PublicMqttQueue>();
 			services.AddSingleton<IRemoteStorageQueue, RemoteStorageQueue>();
